+++ packages/remix-express/CHANGELOG.md
@@ -1,5 +1,26 @@
 # `@remix-run/express`
 
+## 2.12.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@2.12.0`
+
+## 2.11.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@2.11.2`
+
+## 2.11.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@2.11.1`
+
 ## 2.11.0
 
 ### Patch Changes